fib: merge FibEntryDyn into FibEntry
Previously, when both real entry and virtual entry are needed at
FIB StartDepth, information for both is stored in the same
FibEntry struct. Since the virtual entry could be updated
independent from the real entry, in order not to lose counters
during virtual entry update, the counters are placed in a
separate FibEntryDyn struct that is retained during virtual
entry update. This causes overhead in forwarding for every FIB
entry due to extra indirection.

Now, a FibEntry is either a real entry or a virtual entry. When
both are needed, the virtual entry is inserted to the hashtable
and the real entry is accessible via realEntry pointer. Thus,
virtual entry update does not require RCU replacement of the
real entry. This allows counters to be placed back into FibEntry.
Indirection overhead is only paid at FIB StartDepth, not at any
other name length.