ethface: separate port and face creation
Previously, Ethernet port is created automatically when the first face
using that port is created, and port configuration is embedded in the
face locator. This leads to an inconsistent behavior: port configuration
is used when creating the first face but ignored otherwise.

Previously, Ethernet port may switch between RxFlow and RxTable
dispatching logic when creating more faces. The fallback logic is rather
fragile and often results in all the faces to stop working altogether.

Now, Ethernet port creation is moved to a separate 'createEtherPort'
mutation, which must be invoked before trying to create a face. The
choice between RxFlow and RxTable must be made during port creation, and
there is no more automatic fallback logic.

CLI, schemas, documentation are not yet updated in this commit.