ethnetif: AF_PACKET bugfixes
Recover from panic in GraphQL mutation resolver, which is a common source of panics.
However, this would not catch panics in field resolvers, which depends on
https://github.com/graphql-go/graphql/issues/243

Ensure Ethtool instance is created before being used.

Don't create ethdev with AF_PACKET or AF_XDP if the netif is not UP.

In face.md, list the command/mutation in each role that accepts locator.
https://github.com/usnistgov/ndn-dpdk/issues/57

Write a note in face.md on how to disable PACKET_QDISC_BYPASS.
https://github.com/containernet/containernet/issues/232