disk: DiskStore_GetData reply via callback
In forwarder data plane, after loading Data from disk, it's necessary to
re-dispatch the Interest via InputDemux because the NDT could have
changed. Hence, the previous API of replying to ringbuffer does not
match the use case. Now it's a more generic API of a callback function,
which avoids coupling with InputDemux.