iface+tg: allow trafficgen on EthRxTable
iface.ActivateRxGroup and iface.ActivateTxFace can now create the first
RxLoop and TxLoop automatically if one does not exist. Combined with
per-face InputDemuxes, this allows trafficgen to operate on a shared
RxLoop via per-face InputDemuxes. While this is slower than the
recommended setup, it allows running a single instance of trafficgen on
an otherwise incompatible Ethernet port/face.